backend/pyproject.toml
[tool.setuptools.packages.find]
where = [".."]
include = ["backend*"]

[project]
name = "recallmap-backend"
version = "0.1.0"
description = "RecallMap backend — Privacy-First Learning AI Agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    # ingest
    "pymupdf>=1.24.0",          # PDF 解析
    "python-pptx>=0.6.23",      # PPT 解析
    "python-docx>=1.1.0",       # Word 解析
    "pillow>=10.3.0",           # 圖片處理
    # integrations
    "notion-client>=2.2.1",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.127.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
